Split action tables to internal and shared
